niedokonczone zmiany
