+++ package.json
@@ -9,13 +9,8 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-<<<<<<< HEAD
     "bootstrap": "^4.5.0",
     "formik": "^2.1.5",
-=======
-    "formik": "^2.1.5",
-    "bootstrap": "^4.5.0",
->>>>>>> master
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
