+++ src/App.js
@@ -5,6 +5,12 @@ import SignUp from './components/SignUp/SignUp';
 import Home from './components/Home';
 import About from './components/About';
 import Favorite from './components/Challlenges';
+
+
+import ChallengeImage1 from "./components/image/challenge1.jpg";
+import ChallengeImage2 from "./components/image/challenge2.jpg";
+import ChallengeImage3 from "./components/image/challenge3.jpg";
+
 import ChallengeDescription from './components/ChallengeDescription';
 import UserProfile from './components/UserProfile';
 import Diary from './components/Diary';
@@ -43,7 +49,7 @@ const useStyles = makeStyles(theme => ({
 }));
 
 
-function AppContent() {
+function AppContent(props) {
   const classes = useStyles();
   return (
     <BrowserRouter>
@@ -58,10 +64,10 @@ function AppContent() {
               <About />
             </Route>
             <Route exact path="/challenges">
-            <ChallengesList />
+              <ChallengesList challangeUpdater={props.handleChallangeUpdate} multiFilterChange={props.multiFilterChange} challanges={props.challanges} filter={props.filter} categories={props.categories} />
             </Route>
             <Route path="/challenges/:id">
-              <ChallengeDescription />
+              <ChallengeDescription  challangeUpdater={props.handleChallangeUpdate} challanges={props.challanges} categories={props.categories} />
             </Route>
             <Route path="/userProfile">
               <UserProfile />
@@ -70,7 +76,7 @@ function AppContent() {
               <Diary />
             </Route>
             <Route path="/settings">
-              <Settings />
+              <Settings challanges={props.challanges} />
             </Route>
             <Route path="/">
               <Home />
@@ -83,11 +89,170 @@ function AppContent() {
 }
 
 class App extends React.Component {
-  state = {
-    user: null,
-    log: null,
+
+
+
+  constructor(props){
+
+    super(props);
+
+  this.state = {
+    log: Logged,
+    filter: {
+      period: [1,30],
+      text: "",
+      groups: [0,1,2,3]
+    },
+    categories: [
+      {id: 0, title:"Sylwetka", value:5, selected: false},
+      {id: 1, title: "Witalność", value:5, selected: false}, 
+      {id: 2, title: "Zwyczaje", value:5, selected: false}, 
+      {id: 3, title: "Dieta", value:5, selected: false}
+    ],
+    challanges: [
+      {
+        img: ChallengeImage1,
+        title: 'Spacer dla Twojego umysłu.',
+        id: 1,
+        category: 0,
+        period: 5
+      },
+      {
+        img: ChallengeImage2,
+        title: 'Wyśnij sobie spokój.',
+        id: 2,
+        category: 1,
+        period: 25
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 3,
+        category: 2,
+        period: 5
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 4,
+        category: 3,
+        period: 5
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 5,
+        category: 0,
+        period: 5
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 6,
+        category: 1,
+        period: 5
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 7,
+        category: 2,
+        period: 1
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 8,
+        category: 3,
+        period: 1
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 9,
+        category: 3,
+        period: 5
+      },
+      {
+        img: ChallengeImage3,
+        title: 'Pij wodę, będziesz wielki.',
+        id: 0,
+        category: 2,
+        period: 5
+      },
+    ]
+  };
+}
+
+
+handleMultiFilterChange = (grp, event, val) => {
+
+  switch(grp){
+    case "category":
+
+        // 1. Make a shallow copy of the items
+        let categories = [...this.state.categories];
+        // 2. Make a shallow copy of the item you want to mutate
+
+        categories.forEach((category, catIndex) => {
+
+          category.selected = false;
+          
+          val.forEach((value,valueIndex) => {
+          
+            if(value.id === category.id){
+              category.selected = true;
+              categories[catIndex] = category;
+            }
+          });
+        });
+
+        this.setState({categories});
+      break;
+    case "filter_text":
+      
+      this.setState({
+        filter:{
+          text: event
+        }
+      });
+      break;
+
+    case "filter_period":
+      
+      this.setState({
+        filter:{
+          period: val
+        }
+      });
+      break;
+    case "filter_groups":
+      let groups = event;
+
+      let newGroupsArr = [];
+
+      groups.forEach((item, index) => {
+        if(item.selected){
+          newGroupsArr.push(item.key);
+        }
+      });
+
+      this.setState({
+        filter:{
+          groups: newGroupsArr
+        }
+      });
+      break;
+    default:
   }
+}
   
+
+
+  handleChallangeUpdate = () => {
+    console.log("Handle my change in app.js please");
+  }
+
   handleApp= () => {
     this.setState({
       log: Logged
@@ -110,13 +275,13 @@ class App extends React.Component {
    
     switch(this.state.log) {
       case Logged:
-        return <AppContent />
+        return <AppContent challangeUpdater={this.state.handleChallangeUpdate} multiFilterChange={this.state.handleMultiFilterChange} challanges={this.state.challanges} filter={this.state.filter} categories={this.state.categories} />
       case Login:
         return <SignIn onApp={this.handleApp} />
       case Register:
         return <SignUp onLogin={this.handleLogin} onApp={this.handleApp} />
       default:
-        return <OnBoarding onLogin={this.handleLogin} onRegister={this.handleRegister} />  
+        return <OnBoarding onLogin={this.handleLogin} onRegister={this.handleRegister} />
     }
   
   }
